Run tests using Docker
This is a work in progress commit which starts Solr 5
in a local Docker container for the duration of the test suite.

This creates the test core but does not configure it completely to match our previous test config.

Currently we are not starting SolrCloud and those tests break completely because ZooKeeper is not exposed following the instructions at https://hub.docker.com/_/solr/.

This also does not yet enable the Travis CI support for Docker.